fix: same int32 randint overflow in Heisenberg3DSampler (#38)
Heisenberg3DSampler.sample_nearby_vector carried the same
randint(2**32-1) pattern fixed in RegularizedMeanField: NumPy's legacy
RandomState.randint maps a default dtype=int to C long, which is 32-bit
on 64-bit Windows, so the call raises "high is out of bounds for int32"
there.

The seed is passed straight into the njit'd jit_sample_nearby_vector,
which types a Python int as int64 already, so numba sees the same type
and the draw is unchanged: ranges that fit in 32 bits go through the
same 32-bit bounded generator regardless of output dtype.

The class __init__ raises NotImplementedError, so this is unreachable
today; fixed for consistency so reviving the class doesn't reintroduce
the bug.

+++ coniii/experimental/samplers.py
@@ -606,7 +606,7 @@ class Heisenberg3DSampler(Sampler):
             else:
                 ophi = 2*np.pi - np.arccos( v[0]/r )
 
-        return jit_sample_nearby_vector( self.rng.randint(2**32-1),v,nSamples,otheta,ophi,sigma )
+        return jit_sample_nearby_vector( self.rng.randint(2**32-1, dtype=np.int64),v,nSamples,otheta,ophi,sigma )
 
     def _sample_nearby_vector(self, v, nSamples=1, otheta=None, ophi=None, sigma=.1):
         """
